# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Aux API",
    "description": "Backend API for Aux - Universal Playlist Converter. Convert playlists between Spotify and Apple Music seamlessly.",
    "version": "1.0.0",
    "contact": {
      "name": "Aux Support",
      "email": "ayomideadekoya266@gmail.com",
      "url": "https://github.com/elcruzo/aux"
    }
  },
  "servers": [
    {
      "url": "https://aux-50dr.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/convert": {
      "post": {
        "tags": ["Conversion"],
        "summary": "Convert playlist",
        "description": "Convert a playlist from one platform to another",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conversion result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversionResult"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Track": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "description": "Duration in milliseconds"
          },
          "isrc": {
            "type": "string",
            "nullable": true
          },
          "platform": {
            "type": "string",
            "enum": ["spotify", "apple"]
          },
          "uri": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ConversionRequest": {
        "type": "object",
        "required": ["playlistId", "playlistName", "direction"],
        "properties": {
          "playlistId": {
            "type": "string"
          },
          "playlistName": {
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": ["spotify-to-apple", "apple-to-spotify"]
          }
        }
      },
      "ConversionResult": {
        "type": "object",
        "properties": {
          "playlistId": {
            "type": "string"
          },
          "playlistName": {
            "type": "string"
          },
          "totalTracks": {
            "type": "integer"
          },
          "successfulMatches": {
            "type": "integer"
          },
          "failedMatches": {
            "type": "integer"
          },
          "targetPlaylistId": {
            "type": "string",
            "nullable": true
          },
          "targetPlaylistUrl": {
            "type": "string",
            "nullable": true
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackMatch"
            }
          }
        }
      },
      "TrackMatch": {
        "type": "object",
        "properties": {
          "sourceTrack": {
            "$ref": "#/components/schemas/Track"
          },
          "targetTrack": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Track"
              }
            ]
          },
          "confidence": {
            "type": "string",
            "enum": ["high", "medium", "low", "none"]
          },
          "matchType": {
            "type": "string",
            "enum": ["isrc", "search", "manual"],
            "nullable": true
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Please authenticate first"
                }
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}